src/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "learnflow-backend"
version = "0.1.0"
description = "LearnFlow AI Backend – Agentic Python Services"
readme = "README.md"
requires-python = ">=3.11"

authors = [
  { name = "LearnFlow Team" }
]

dependencies = [
  "fastapi>=0.110,<0.120",
  "uvicorn[standard]>=0.29,<0.31",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "tenacity>=8.2",

  # Async / concurrency
  "anyio>=4.3,<5.0",

  # Observability
  "structlog>=24.1",

  # AI / agents (add as needed)
  "openai>=1.14,<2.0",
  "langchain-core>=0.1.40,<0.3.0",

  # Utilities
  "python-dotenv>=1.0",
  "aiokafka>=0.10,<0.12",
  "dapr>=1.13,<1.15"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.9",
  "types-python-dateutil"
]



# TOOLING CONFIGURATION

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]



# HATCH (PACKAGING)

[tool.hatch.build.targets.wheel]
packages = ["app"]

